+++ pages/dashboard.js
@@ -109,7 +109,6 @@ class Dashboard extends Component {
         <NotLoggedIn message='Log in with your OSM account to see your personalized dashboard' />
       )
     }
-
     // Dashboard header variables
     let profileImage = null
     let name = null
@@ -131,6 +130,7 @@ class Dashboard extends Component {
       accountId = authenticatedUser.account.id
     }
     const recordsExport = [{ ...account.records, badgeCount, campaignCount, name }]
+    const userHasEdits = edit_times.length > 0
     const userHasCampaigns = (favorites.length > 0 || assignments.length > 0 || campaignCount > 0)
     return (
       <div className='dashboard'>
@@ -155,43 +155,31 @@ class Dashboard extends Component {
         <div id='blurb' className='row'>
           <DashboardBlurb {...osmesaData} project={project} />
           <div className='widget-33 page-actions'>
-            {
-              edit_times > 0
-                ? <CSVExport filename={`${name}_ScoreboardData.csv`} data={recordsExport} />
-                : <></>
+            {userHasEdits &&
+              <CSVExport filename={`${name}_ScoreboardData.csv`} data={recordsExport} />
             }
           </div>
         </div>
         {
-          (userHasCampaigns || isAdmin(authenticatedUser.account.roles))
-            ? <section id='admin&assignments' className='section--dark'>
-              <div className='row'>
-                {isAdmin(authenticatedUser.account.roles) && this.renderAdmin()}
-                {
-                  (favorites.length > 0 || assignments.length > 0 || campaignCount > 1)
-                    ? <DashboardAssignments
-                      favorites={favorites}
-                      assignments={assignments}
-                      authenticatedUser={authenticatedUser}
-                      all={allCampaigns}
-                    />
-                    : <></>
-                }
-              </div>
-            </section>
-            : <></>
+          (userHasCampaigns || isAdmin(authenticatedUser.account.roles)) &&
+          <section id='admin-assignments' className='section--dark'>
+            <div className='row'>
+              {isAdmin(authenticatedUser.account.roles) && this.renderAdmin()}
+              {userHasCampaigns && this.renderAssignments()}
+            </div>
+          </section>
         }
         {
-          edit_times > 0
+          userHasEdits
             ? <>
-              <section id='map'>
+              <section id='dashboard_map'>
                 <div className='row'>
                   <div className='map-lg'>
                     <UserExtentMap uid={uid} extent={extent_uri} />
                   </div>
                 </div>
               </section>
-              <section id='charts'>
+              <section id='dashboard_charts'>
                 <div className='row'>
                   <div className='widget-container'>
                     <div className='widget-66'>
@@ -207,7 +195,7 @@ class Dashboard extends Component {
                   </div>
                 </div>
               </section>
-              <section id='sidebar-badges-heatmap'>
+              <section id='dashboard_sidebar-badges-heatmap'>
                 <div className='row widget-container'>
                   <DashboardSidebar teams={teams} osmesaData={osmesaData} loggedIn={loggedIn} />
                   <div className='widget-75'>
@@ -218,10 +206,8 @@ class Dashboard extends Component {
                   <CalendarHeatmap times={edit_times} />
                 </div>
               </section>
-        </>
-            : <section className='row'>
-              <img className='img--blank-state' src='static/dashboard-temp/open_maps.svg' />
-            </section>
+              </>
+            : this.renderBlankState()
         }
       </div>
     )
@@ -240,6 +226,27 @@ class Dashboard extends Component {
         <br />
         <br />
       </div>
+    )
+  }
+
+  renderAssignments () {
+    const { authenticatedUser } = this.props
+    const { assignments, favorites, allCampaigns } = authenticatedUser.account
+    return (
+      <DashboardAssignments
+        favorites={favorites}
+        assignments={assignments}
+        authenticatedUser={authenticatedUser}
+        all={allCampaigns}
+      />
+    )
+  }
+
+  renderBlankState () {
+    return (
+      <section className='row'>
+        <img className='img--blank-state' src='static/dashboard-temp/open_maps.svg' />
+      </section>
     )
   }
 }